Unify some test dsl
Moves `file_context` into a shared space because we use it in both the
unit tests and integration tests.
